[Edit] rating 데이터 타입 Int -> Double 변경, StarRating color 변경

+++ FinalHrmiProjects/View/Record/AddTagView.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct DrinkTag: Identifiable, Hashable {
     var id = UUID()
     var name: String
-    var rating: Int
+    var rating: Double
 }
 
 struct AddTagView: View {
